+++ src/Attributes/Services/AttributesParser.php
@@ -17,25 +17,44 @@ use Bitrix24\SDK\Attributes\ApiBatchMethodMetadata;
 use Bitrix24\SDK\Attributes\ApiBatchServiceMetadata;
 use Bitrix24\SDK\Attributes\ApiEndpointMetadata;
 use Bitrix24\SDK\Attributes\ApiServiceMetadata;
+use Bitrix24\SDK\Core\Credentials\Scope;
+use Bitrix24\SDK\Core\Exceptions\UnknownScopeCodeException;
 use ReflectionClass;
 use Symfony\Component\Filesystem\Filesystem;
 use Typhoon\Reflection\TyphoonReflector;
+
 use function Typhoon\Type\stringify;
 
 readonly class AttributesParser
 {
     public function __construct(
         private TyphoonReflector $typhoonReflector,
-        private Filesystem       $filesystem,
-    )
-    {
+        private Filesystem $filesystem,
+    ) {
     }
 
     /**
      * @param class-string[] $sdkClassNames
-     * @return array<string, mixed>
+     * @param non-empty-string $sdkBaseDir
+     * @return array<string, array{
+     *    sdk_scope: string,
+     *    name: string,
+     *    documentation_url: string|null,
+     *    description: string|null,
+     *    is_deprecated: bool,
+     *    deprecation_message: string|null,
+     *    sdk_method_name: string,
+     *    sdk_method_file_name: string,
+     *    sdk_method_file_start_line: int,
+     *    sdk_method_file_end_line: int,
+     *    sdk_class_name: class-string,
+     *    sdk_return_type_class: string|null,
+     *    sdk_return_type_file_name: string|null
+     * }>
+     *
+     * @throws UnknownScopeCodeException
      */
-    public function getSupportedInSdkApiMethods(array $sdkClassNames, string $sdkBaseDir): array
+    public function getSupportedInSdkApiMethods(array $sdkClassNames, string $sdkBaseDir, ?Scope $scope = null): array
     {
         $supportedInSdkMethods = [];
         foreach ($sdkClassNames as $className) {
@@ -66,19 +85,28 @@ readonly class AttributesParser
                         $returnTypeName = $method->getReturnType()->getName();
                         if (class_exists($returnTypeName)) {
                             $reflectionReturnType = new ReflectionClass($returnTypeName);
-                            $returnTypeFileName = substr($this->filesystem->makePathRelative($reflectionReturnType->getFileName(), $sdkBaseDir), 0, -1);
+                            $returnTypeFileName = substr(
+                                $this->filesystem->makePathRelative($reflectionReturnType->getFileName(), $sdkBaseDir),
+                                0,
+                                -1
+                            );
                         }
                     }
 
                     $supportedInSdkMethods[$instance->name] = [
-                        'sdk_scope' => $apiServiceAttrInstance->scope->getScopeCodes() === [] ? '' : $apiServiceAttrInstance->scope->getScopeCodes()[0],
+                        'sdk_scope' => $apiServiceAttrInstance->scope->getScopeCodes(
+                        ) === [] ? '' : $apiServiceAttrInstance->scope->getScopeCodes()[0],
                         'name' => $instance->name,
                         'documentation_url' => $instance->documentationUrl,
                         'description' => $instance->description,
                         'is_deprecated' => $instance->isDeprecated,
                         'deprecation_message' => $instance->deprecationMessage,
                         'sdk_method_name' => $method->getName(),
-                        'sdk_method_file_name' => substr($this->filesystem->makePathRelative($method->getFileName(), $sdkBaseDir), 0, -1),
+                        'sdk_method_file_name' => substr(
+                            $this->filesystem->makePathRelative($method->getFileName(), $sdkBaseDir),
+                            0,
+                            -1
+                        ),
                         'sdk_method_file_start_line' => $method->getStartLine(),
                         'sdk_method_file_end_line' => $method->getEndLine(),
                         'sdk_class_name' => $className,
@@ -89,6 +117,21 @@ readonly class AttributesParser
                 }
             }
         }
+
+        if ($scope instanceof Scope) {
+            $allMethods = $supportedInSdkMethods;
+            $supportedInSdkMethods = [];
+            foreach ($allMethods as $method) {
+                // skip methods without scope
+                if ($method['sdk_scope'] === '') {
+                    continue;
+                }
+                if ($scope->contains($method['sdk_scope'])) {
+                    $supportedInSdkMethods[] = $method;
+                }
+            }
+        }
+
         return $supportedInSdkMethods;
     }
 
@@ -124,7 +167,9 @@ readonly class AttributesParser
                     $sdkReturnTypeTyphoon = null;
                     if ($method->getReturnType() !== null) {
                         // get return type from phpdoc annotation
-                        $sdkReturnTypeTyphoon = stringify($typhoonClassMeta->methods()[$method->getName()]->returnType());
+                        $sdkReturnTypeTyphoon = stringify(
+                            $typhoonClassMeta->methods()[$method->getName()]->returnType()
+                        );
                     }
 
                     $supportedInSdkMethods[$instance->name][] = [